Refactor api/ and ~contribute
