+++ lighthouse/live.txt
@@ -1,5 +1,5 @@
 https://hollowverse.com/elon-musk
-https://hollowverse.com/~/contribute?name=Meat+Loaf&slug=meat-loaf
+https://hollowverse.com/~contribute?name=Meat+Loaf&slug=meat-loaf
 https://hollowverse.com/martin-freeman
 https://hollowverse.com/
 https://hollowverse.com/lucille-ball
