+++ pages/api/content-change-notify.ts
@@ -1,1 +1,1 @@
-export { contentChangeNotify as default } from '~/lib/api/contentChangeNotify';
+export { contentChangeNotify as default } from '~/lib/api/contentChangeNotify/contentChangeNotify';
